% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{counter}
\usage{
counter(dat, var)
}
\arguments{
\item{dat}{The dataset}

\item{var}{A categorical variable}
}
\value{
A table of counts for each of the categories of the categorical variable.
}
\description{
Generates a table of counts, for any given dataset and categorical variable.
}
\examples{
my_tab <- counter(infert, "case")
}
