% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumla.R
\name{jumla}
\alias{jumla}
\title{jumla
takes in two numbers, and adds them.}
\usage{
jumla(x, y)
}
\arguments{
\item{x}{The first argument}

\item{y}{The second argument}
}
\value{
a value that is the summation of x and y
}
\description{
jumla
takes in two numbers, and adds them.
}
\examples{
summation <- jumla(10,20)

}
