% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_attendees.R
\docType{data}
\name{meetup_attendees}
\alias{meetup_attendees}
\title{meetup_attendees}
\format{
A data frame with 3 variables:
\describe{
\item{\code{first_name}}{First name}
\item{\code{last_name}}{Last _name}
\item{\code{country}}{Country where the attendee is based}
}
}
\usage{
data(meetup_attendees)
}
\description{
This data contains information about some attendees of the meetup
}
\keyword{datasets}
